Create EDA on Nipah outbreak in Kerala
This is a real life application project using pandas and matplotlib. The code is  written in Python. This project is based on the  'Nipah  virus ' outbreak in Kerala . In this project , the number of cases reported and the count of deaths are shown. 2018- present data is used for the same.